requirement.txt
alpaca-py
alpaca-trade-api
pandas
motor
pymongo
python-dotenv
fastapi
uvicorn
pytz
ollama
colorama